refactor: Member 패키지 변경
