+++ backend/baton/src/main/java/touch/baton/domain/member/vo/Email.java
@@ -1,4 +1,4 @@
-package touch.baton.member.domain.vo;
+package touch.baton.domain.member.vo;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -10,12 +10,12 @@ import static lombok.AccessLevel.PROTECTED;
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 @Embeddable
-public class Name {
+public class Email {
 
-    @Column(name = "name")
+    @Column(name = "email")
     private String value;
 
-    public Name(final String value) {
+    public Email(final String value) {
         this.value = value;
     }
 }
